feat: 강의 조회 서비스

+++ src/main/kotlin/com/mkroo/lmsdemo/dto/LectureOpeningRequest.kt
@@ -2,6 +2,6 @@ package com.mkroo.lmsdemo.dto
 
 data class LectureOpeningRequest(
     val title: String,
-    val maxStudentCount: Int,
-    val price: Int
+    val maxStudentCount: Long,
+    val price: Long
 )
